+++ NEWS.md
@@ -1,4 +1,6 @@
 # driveR (development version)
+## Major Changes
+- added support for gene-level SCNA data frames to be used as input for `create_features_df()` and added an example gene-level SCNA data frame `example_gene_scna_table`
 
 # driveR 0.4.1